third_party/WebKit/LayoutTests/animations/responsive/animations-responsive-columnWidth.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<div id='container'>
<div id='element'></div>
</div>

<script>
var element = document.getElementById('element');
var container = document.getElementById('container');

test(function() {
    element.style.fontSize = '10px';
    var player = element.animate([{columnWidth: '3em'}, {columnWidth: '5em'}], 10);
    player.pause();
    player.currentTime = 5;
    element.style.fontSize = '20px';
    assert_equals(getComputedStyle(element).columnWidth, '80px');
}, 'column-width responsive to style changes');

test(function() {
    element.style.fontSize = '10px';
    var player = element.animate([{columnWidth: '40px'}, {columnWidth: 'calc(40px - 2em)'}], 10);
    player.pause();

    player.currentTime = 5;
    element.style.fontSize = '40px';
    assert_equals(getComputedStyle(element).columnWidth, '0px');

    player.currentTime = 7.5;
    assert_equals(getComputedStyle(element).columnWidth, '0px');
}, 'column-width clamped to 0px');

test(function() {
    container.style.columnWidth = 'auto';
    var player = element.animate([{columnWidth: 'inherit'}, {columnWidth: '20px'}], 4000);
    player.pause();

    player.currentTime = 1000;
    assert_equals(getComputedStyle(element).columnWidth, 'auto');

    container.style.columnWidth = '100px';
    assert_equals(getComputedStyle(element).columnWidth, '80px');

    container.style.columnWidth = 'auto';
    assert_equals(getComputedStyle(element).columnWidth, 'auto');
}, 'column-width responsive to inherited changes');
</script>
